refactor: move parseRules from VariableNameFilter to NameFilterRule
parseRules converts raw configuration strings into NameFilterRule objects,
making it a natural static factory on NameFilterRule rather than a utility
buried in VariableNameFilter. This removes the implicit dependency of
DefaultRecordFilter on VariableNameFilter internals.
