+++ zeebe/exporter-filter/src/main/java/io/camunda/zeebe/exporter/filter/VariableNameFilter.java
@@ -10,9 +10,7 @@ package io.camunda.zeebe.exporter.filter;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.value.VariableRecordValue;
 import io.camunda.zeebe.util.SemanticVersion;
-import java.util.Collections;
 import java.util.List;
-import java.util.Objects;
 
 public final class VariableNameFilter implements ExporterRecordFilter, RecordVersionFilter {
 
@@ -33,21 +31,6 @@ public final class VariableNameFilter implements ExporterRecordFilter, RecordVer
     }
 
     return nameFilter.accept(variableRecordValue.getName());
-  }
-
-  public static List<NameFilterRule> parseRules(
-      final List<String> rawList, final NameFilterRule.Type type) {
-
-    if (rawList == null || rawList.isEmpty()) {
-      return Collections.emptyList();
-    }
-
-    return rawList.stream()
-        .filter(Objects::nonNull)
-        .map(String::trim)
-        .filter(s -> !s.isEmpty())
-        .map(pattern -> new NameFilterRule(type, pattern))
-        .toList();
   }
 
   @Override
